feat: :sparkles: CreateTransaction
continues refactoring business and infrastructure code to segregated layers
